+++ lib/slimmer/rspec.rb
@@ -2,10 +2,10 @@ require 'rspec/core'
 
 require 'slimmer'
 require 'slimmer/test'
-require 'slimmer/test_helpers/shared_templates'
+require 'slimmer/test_helpers/govuk_components'
 
 RSpec.configure do |config|
-  config.include Slimmer::TestHelpers::SharedTemplates
+  config.include Slimmer::TestHelpers::GovukComponents
 
   config.before { stub_shared_component_locales }
 end